+++ Assets/Scripts/Vanilla/Frameworks/RandomChinaEvents/RandomChinaEventDefinition.cs
@@ -9,15 +9,17 @@ namespace MVZ2.Vanilla.RandomChina
 {
     public abstract class RandomChinaEventDefinition : Definition
     {
-        public RandomChinaEventDefinition(string nsp, string path, string name, float weight = 1) : base(nsp, path)
+        public RandomChinaEventDefinition(string nsp, string path, string name, string description, float weight = 1) : base(nsp, path)
         {
-            Text = name;
+            EventName = name;
+            EventDescription = description;
             Weight = weight;
         }
         public abstract void Run(Entity contraption, RandomGenerator rng);
 
         public override string GetDefinitionType() => VanillaDefinitionTypes.RANDOM_CHINA_EVENT;
-        public string Text { get; set; }
+        public string EventName { get; set; }
+        public string EventDescription { get; set; }
         public float Weight { get; set; }
     }
 }