+++ Assets/Scripts/Vanilla/GameContent/RandomChinaEvents/AncientEgyptEvent.cs
@@ -18,7 +18,7 @@ namespace MVZ2.GameContent.RandomChinaEvents
     [RandomChinaEventDefinition(VanillaRandomChinaEventNames.ancientEgypt)]
     public class AncientEgyptEvent : RandomChinaEventDefinition
     {
-        public AncientEgyptEvent(string nsp, string path) : base(nsp, path, NAME)
+        public AncientEgyptEvent(string nsp, string path) : base(nsp, path, NAME, DESCRIPTION)
         {
         }
         public override void Run(Entity contraption, RandomGenerator rng)
@@ -36,5 +36,7 @@ namespace MVZ2.GameContent.RandomChinaEvents
         }
         [TranslateMsg("随机瓷器事件名称", VanillaStrings.CONTEXT_RANDOM_CHINA_EVENT_NAME)]
         public const string NAME = "神秘埃及";
+        [TranslateMsg("随机瓷器事件描述", VanillaStrings.CONTEXT_RANDOM_CHINA_EVENT_DESCRIPTION)]
+        public const string DESCRIPTION = "每行召唤一个木乃伊，然后……";
     }
 }