+++ Assets/Scripts/Vanilla/GameContent/RandomChinaEvents/IndependenceDayEvent.cs
@@ -15,7 +15,7 @@ namespace MVZ2.GameContent.RandomChinaEvents
     [RandomChinaEventDefinition(VanillaRandomChinaEventNames.independenceDay)]
     public class IndependenceDayEvent : RandomChinaEventDefinition
     {
-        public IndependenceDayEvent(string nsp, string path) : base(nsp, path, NAME)
+        public IndependenceDayEvent(string nsp, string path) : base(nsp, path, NAME, DESCRIPTION)
         {
         }
         public override void Run(Entity contraption, RandomGenerator rng)
@@ -38,6 +38,8 @@ namespace MVZ2.GameContent.RandomChinaEvents
         }
         [TranslateMsg("随机瓷器事件名称", VanillaStrings.CONTEXT_RANDOM_CHINA_EVENT_NAME)]
         public const string NAME = "独立日";
+        [TranslateMsg("随机瓷器事件描述", VanillaStrings.CONTEXT_RANDOM_CHINA_EVENT_DESCRIPTION)]
+        public const string DESCRIPTION = "生成10个红色不死飞行物";
         public const int SPAWN_COUNT = 10;
     }
 }