+++ Assets/Scripts/Vanilla/GameContent/RandomChinaEvents/Pokers/AceOfSpadesEvent.cs
@@ -11,7 +11,7 @@ namespace MVZ2.GameContent.RandomChinaEvents
     [RandomChinaEventDefinition(VanillaRandomChinaEventNames.aceOfSpades)]
     public class AceOfSpadesEvent : AceAbstractEvent
     {
-        public AceOfSpadesEvent(string nsp, string path) : base(nsp, path, NAME)
+        public AceOfSpadesEvent(string nsp, string path) : base(nsp, path, NAME, DESCRIPTION)
         {
         }
         protected override void Transform(Entity target, Entity china)
@@ -20,5 +20,7 @@ namespace MVZ2.GameContent.RandomChinaEvents
         }
         [TranslateMsg("随机瓷器事件名称", VanillaStrings.CONTEXT_RANDOM_CHINA_EVENT_NAME)]
         public const string NAME = "黑桃A";
+        [TranslateMsg("随机瓷器事件描述", VanillaStrings.CONTEXT_RANDOM_CHINA_EVENT_DESCRIPTION)]
+        public const string DESCRIPTION = "将所有怪物和掉落物变为星之碎片";
     }
 }