+++ Assets/Scripts/Vanilla/GameContent/RandomChinaEvents/RedstoneReadyEvent.cs
@@ -14,7 +14,7 @@ namespace MVZ2.GameContent.RandomChinaEvents
     [RandomChinaEventDefinition(VanillaRandomChinaEventNames.redstoneReady)]
     public class RedstoneReadyEvent : RandomChinaEventDefinition
     {
-        public RedstoneReadyEvent(string nsp, string path) : base(nsp, path, NAME)
+        public RedstoneReadyEvent(string nsp, string path) : base(nsp, path, NAME, DESCRIPTION)
         {
         }
         public override void Run(Entity contraption, RandomGenerator rng)
@@ -40,5 +40,7 @@ namespace MVZ2.GameContent.RandomChinaEvents
         }
         [TranslateMsg("随机瓷器事件名称", VanillaStrings.CONTEXT_RANDOM_CHINA_EVENT_NAME)]
         public const string NAME = "红石俱备";
+        [TranslateMsg("随机瓷器事件描述", VanillaStrings.CONTEXT_RANDOM_CHINA_EVENT_DESCRIPTION)]
+        public const string DESCRIPTION = "生成20个红石";
     }
 }