+++ Assets/Scripts/Vanilla/GameContent/RandomChinaEvents/Tarots/TheHangedManEvent.cs
@@ -15,7 +15,7 @@ namespace MVZ2.GameContent.RandomChinaEvents
     [RandomChinaEventDefinition(VanillaRandomChinaEventNames.theHangedMan)]
     public class TheHangedManEvent : RandomChinaEventDefinition
     {
-        public TheHangedManEvent(string nsp, string path) : base(nsp, path, NAME)
+        public TheHangedManEvent(string nsp, string path) : base(nsp, path, NAME, DESCRIPTION)
         {
         }
         public override void Run(Entity contraption, RandomGenerator rng)
@@ -32,5 +32,7 @@ namespace MVZ2.GameContent.RandomChinaEvents
         }
         [TranslateMsg("随机瓷器事件名称", VanillaStrings.CONTEXT_RANDOM_CHINA_EVENT_NAME)]
         public const string NAME = "XII-倒吊人";
+        [TranslateMsg("随机瓷器事件描述", VanillaStrings.CONTEXT_RANDOM_CHINA_EVENT_DESCRIPTION)]
+        public const string DESCRIPTION = "生成6个反则卫星";
     }
 }