+++ Assets/Scripts/Vanilla/GameContent/RandomChinaEvents/Tarots/TheStarEvent.cs
@@ -15,7 +15,7 @@ namespace MVZ2.GameContent.RandomChinaEvents
     [RandomChinaEventDefinition(VanillaRandomChinaEventNames.theStar)]
     public class TheStarEvent : RandomChinaEventDefinition
     {
-        public TheStarEvent(string nsp, string path) : base(nsp, path, NAME)
+        public TheStarEvent(string nsp, string path) : base(nsp, path, NAME, DESCRIPTION)
         {
         }
         public override void Run(Entity contraption, RandomGenerator rng)
@@ -31,5 +31,7 @@ namespace MVZ2.GameContent.RandomChinaEvents
         }
         [TranslateMsg("随机瓷器事件名称", VanillaStrings.CONTEXT_RANDOM_CHINA_EVENT_NAME)]
         public const string NAME = "XVII-星星";
+        [TranslateMsg("随机瓷器事件描述", VanillaStrings.CONTEXT_RANDOM_CHINA_EVENT_DESCRIPTION)]
+        public const string DESCRIPTION = "召唤10枚陨石";
     }
 }