+++ Assets/Scripts/Vanilla/GameContent/RandomChinaEvents/Tarots/TheTowerEvent.cs
@@ -16,7 +16,7 @@ namespace MVZ2.GameContent.RandomChinaEvents
     [RandomChinaEventDefinition(VanillaRandomChinaEventNames.theTower)]
     public class TheTowerEvent : RandomChinaEventDefinition
     {
-        public TheTowerEvent(string nsp, string path) : base(nsp, path, NAME)
+        public TheTowerEvent(string nsp, string path) : base(nsp, path, NAME, DESCRIPTION)
         {
         }
         public override void Run(Entity contraption, RandomGenerator rng)
@@ -33,5 +33,7 @@ namespace MVZ2.GameContent.RandomChinaEvents
         }
         [TranslateMsg("随机瓷器事件名称", VanillaStrings.CONTEXT_RANDOM_CHINA_EVENT_NAME)]
         public const string NAME = "XVI-塔";
+        [TranslateMsg("随机瓷器事件描述", VanillaStrings.CONTEXT_RANDOM_CHINA_EVENT_DESCRIPTION)]
+        public const string DESCRIPTION = "在随机位置掉落16个TNT";
     }
 }