+++ Assets/Scripts/Vanilla/GameContent/RandomChinaEvents/WorldwideCelebrationEvent.cs
@@ -16,7 +16,7 @@ namespace MVZ2.GameContent.RandomChinaEvents
     [RandomChinaEventDefinition(VanillaRandomChinaEventNames.worldwideCelebration)]
     public class WorldwideCelebrationEvent : RandomChinaEventDefinition
     {
-        public WorldwideCelebrationEvent(string nsp, string path) : base(nsp, path, NAME)
+        public WorldwideCelebrationEvent(string nsp, string path) : base(nsp, path, NAME, DESCRIPTION)
         {
         }
         public override void Run(Entity contraption, RandomGenerator rng)
@@ -31,5 +31,7 @@ namespace MVZ2.GameContent.RandomChinaEvents
         }
         [TranslateMsg("随机瓷器事件名称", VanillaStrings.CONTEXT_RANDOM_CHINA_EVENT_NAME)]
         public const string NAME = "普天同庆";
+        [TranslateMsg("随机瓷器事件描述", VanillaStrings.CONTEXT_RANDOM_CHINA_EVENT_DESCRIPTION)]
+        public const string DESCRIPTION = "30秒内所有单位的移速、攻击速度和生产速度翻倍";
     }
 }