+++ Assets/Scripts/Vanilla/GameContent/RandomChinaEvents/WrathOfTheSmallEvent.cs
@@ -17,7 +17,7 @@ namespace MVZ2.GameContent.RandomChinaEvents
     [RandomChinaEventDefinition(VanillaRandomChinaEventNames.wrathOfTheSmall)]
     public class WrathOfTheSmallEvent : RandomChinaEventDefinition
     {
-        public WrathOfTheSmallEvent(string nsp, string path) : base(nsp, path, NAME)
+        public WrathOfTheSmallEvent(string nsp, string path) : base(nsp, path, NAME, DESCRIPTION)
         {
         }
         public override void Run(Entity contraption, RandomGenerator rng)
@@ -36,5 +36,7 @@ namespace MVZ2.GameContent.RandomChinaEvents
         }
         [TranslateMsg("随机瓷器事件名称", VanillaStrings.CONTEXT_RANDOM_CHINA_EVENT_NAME)]
         public const string NAME = "小型之怒";
+        [TranslateMsg("随机瓷器事件描述", VanillaStrings.CONTEXT_RANDOM_CHINA_EVENT_DESCRIPTION)]
+        public const string DESCRIPTION = "在随机一行的最左侧生成一个大雪球";
     }
 }